refactor user forms components + validate social networks fields
