+++ frontend/src/locales/en.json
@@ -711,7 +711,8 @@
   "user.gender.male": "Male",
   "user.gender.preferNotToSay": "Prefer not to say",
   "user.gender.selfDescribe": "Prefer to self-describe:",
-  "user.slack": "Username on HOT Slack",
+  "user.slack": "Username on {org} Slack",
+  "user.personalInfo.error": "Type only your username, not the URL.",
   "user.form.save": "Save",
   "user.settings.title": "Settings",
   "user.notifications.title": "Notifications",
