+++ frontend/src/views/settings.js
@@ -2,14 +2,12 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { Redirect } from '@reach/router';
 
-import { UserTopBar } from '../components/user/settings';
+import { UserTopBar } from '../components/user/topBar';
 import { OSMCard, APIKeyCard } from '../components/user/content';
-import {
-  UserInformationForm,
-  UserInterestsCard,
-  UserSettingsForm,
-  UserNotificationsForm,
-} from '../components/user/forms';
+import { PersonalInformationForm } from '../components/user/forms/personalInformation';
+import { UserSettingsForm } from '../components/user/forms/settings';
+import { UserNotificationsForm } from '../components/user/forms/notifications';
+import { UserInterestsForm } from '../components/user/forms/interests';
 import { useSetTitleTag } from '../hooks/UseMetaTags';
 
 export function Settings() {
@@ -23,13 +21,13 @@ export function Settings() {
         <UserTopBar />
         <div className="cf pa4 bg-tan">
           <div className="fl w-100 w-60-l pb3 pr3-l">
-            <UserInterestsCard />
+            <UserInterestsForm />
             <UserSettingsForm />
             <UserNotificationsForm />
             {userDetails.isExpert && <APIKeyCard token={token} />}
           </div>
           <div className="fl w-100 w-40-l pb3 pl3-l">
-            <UserInformationForm />
+            <PersonalInformationForm />
             <OSMCard username={userDetails.username} />
           </div>
         </div>
